school_sys/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
ansible==13.1.0
ansible-core==2.20.1
arandr==0.1.11
argcomplete==3.6.2
asgiref==3.10.0
attrs==25.4.0
autocommand==2.2.2
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
boolean.py==5.0
Brlapi==0.8.7
btrfsutil==6.17.1
build==1.3.0
CacheControl==0.14.4
cachy==0.3.0
cairocffi==1.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.2.1
configobj==5.0.9
crashtest==0.4.1
cryptography==46.0.3
cupshelpers==1.0
dbus-python==1.4.0
dbus_next==0.2.3
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
Django==5.1.15
django-filter==25.2
djangorestframework==3.16.1
dnspython==2.8.0
docopt==0.6.2
dulwich==0.24.10
Faker==39.0.0
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.1
Flask==3.1.2
fluidity-sm==0.2.1
frozenlist==1.8.0
guake @ file:///build/guake/src/guake
gufw==26.4.0
html5lib==1.1
httplib2==0.31.0
idna==3.11
inflect==7.5.0
installer==0.7.0
invoke==2.2.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jeepney==0.9.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.7.0
keyutils==0.6
lark==1.3.1
lensfun==0.3.4
lexicon==3.0.0
libfdt==1.7.2
librt==0.7.3
libtorrent==2.0.11
license-expression==30.4.4
lockfile==0.12.2
louis==3.36.0
lxml==6.0.2
MarkupSafe==3.0.2
menulibre==2.4.0
more-itertools==10.8.0
msgpack==1.1.2
mugshot==0.4.3
multidict==6.7.0
mypy==1.19.0
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
mysqlclient==2.2.7
mysqlx-connector-python==9.5.0
namcap==3.6.0
netifaces==0.11.0
netsnmp-python==1.0a1
nftables==0.1
numpy==2.4.0
OpenFreebuds @ file:///home/hectorio23/.cache/yay/openfreebuds/src/openfreebuds-0.16.1/dist/openfreebuds-0.16.1-py3-none-any.whl#sha256=c6a9fd69cc63d4e2928f15a5e2f474e4925ccd7cc378adc859030514b62b2321
orjson==3.11.5
packaging==25.0
paramiko==4.0.0
pathspec==0.12.1
pbr==7.0.3
pbs-installer==2025.12.17
pexpect==4.9.0
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.5.1
poetry==2.2.1
poetry-core==2.2.1
poetry-plugin-export==1.9.0
propcache==0.4.1
psutil==7.1.3
ptyprocess==0.7.0
pwquality==1.4.5
pyalpm==0.10.12
pyasyncore==1.0.4
pycairo==1.29.0
pycparser==2.23
pycups==2.0.4
pycurl==7.45.7
pyelftools==0.32
PyGObject==3.54.5
PyMySQL==1.1.2
PyNaCl==1.6.1
pyparsing==3.3.1
pyparted==3.13.0
pyproject_hooks==1.2.0
PyQt6==6.10.1
PyQt6_sip==13.10.3
pytz==2025.2
pywal==3.3.0
pywayland==0.4.18
pywlroots==0.17.0
pyxdg==0.28
PyYAML==6.0.3
qasync==0.28.0
QTermWidget==2.3.0
qtile==0.33.1.dev0+g9c51897f.d20250729
ranger-fm==1.9.4
RapidFuzz==3.14.3
referencing==0.37.0
Reflector==2023.6.28.0.36.1
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
resolvelib==1.2.1
rpds-py==0.30.0
scour==0.38.2
SecretStorage==3.3.3
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smbus==1.1
soupsieve==2.8.1
speedtest-cli==2.1.4b1
sqlparse==0.5.3
systemd-python==235
TBB==0.2
tomlkit==0.13.3
trove-classifiers==2025.12.1.14
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.2.post0
udiskie==2.5.8
ufw==0.36.2
Unidecode==1.4.0
urllib3==2.6.2
validate==5.0.9
variety==0.8.13
virtualenv==20.35.4
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wxPython==4.2.3
xcffib==1.6.1
xkbcommon==1.5.1
yarl==1.22.0
yt-dlp==2025.12.8
